Game now takes 'lizard' and 'spock' as choices for the player and computer in the game. The game also takes short-form aliases to save the user from typing out each choice in full: 'r', 'p', 'sc', 'l', 'sp'.
New constant ALIASES houses the short-form alias inputs that a user can type with each index correlating to the index of the VALID_CHOICES. This way the newly made  method can take an alias and convert it to the full word, allowing it to be printed in the console and used for evaluation. An extra  condition can check if the user's input is an alias or not.